+++ sample/src/main/java/com/getyourguide/uitestglazesample/jetpackcomposeviews/JetpackComposeViewsFragment.kt
@@ -4,8 +4,22 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.semantics.testTagsAsResourceId
+import androidx.compose.ui.unit.dp
 import androidx.fragment.app.Fragment
 
 class JetpackComposeViewsFragment : Fragment() {
@@ -17,7 +31,27 @@ class JetpackComposeViewsFragment : Fragment() {
     ): View {
         return ComposeView(requireContext()).apply {
             setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-            setContent { }
+            setContent {
+                UserInterface()
+            }
         }
+    }
+}
+
+@OptIn(ExperimentalComposeUiApi::class, ExperimentalMaterial3Api::class)
+@Composable
+fun UserInterface() {
+    val textEntered = remember { mutableStateOf("") }
+    Column(modifier = Modifier
+        .padding(16.dp)
+        .semantics {
+            testTagsAsResourceId = true
+        }) {
+        Text("Hello World", modifier = Modifier.testTag("hello_world_tag"))
+        TextField(
+            value = textEntered.value,
+            onValueChange = { textEntered.value = it },
+            modifier = Modifier.testTag("text_field_tag")
+        )
     }
 }